refactor(types): split page molecule schemas into modules
