+++ packages/types/src/page/molecules/index.ts
@@ -1,0 +1,14 @@
+export * from "./announcement-bar";
+export * from "./value-props";
+export * from "./reviews-carousel";
+export * from "./contact-form";
+export * from "./newsletter-signup";
+export * from "./search-bar";
+export * from "./map-block";
+export * from "./video-block";
+export * from "./faq-block";
+export * from "./countdown-timer";
+export * from "./social-links";
+export * from "./social-feed";
+export * from "./social-proof";
+
